pyproject.toml
[tool.poetry]
name = "burst_explorer"
version = "1.6.0"
description = "BurstCoin Blockchain Explorer"
authors = ["Lev Lybin <lev.lybin@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
celery = {version = "4.4.1", extras = ["redis"]}
django = "3.0.4"
django-cache-memoize = "0.1.6"
django-cors-headers = "3.2.1"
django-redis = "4.11.0"
djangorestframework = "3.11.0"
jsonschema = "3.2.0"
mysqlclient = "1.4.6"
pycoingecko = "1.2.0"
python-dotenv = "0.12.0"
python-redis-lock = {version = "3.5.0", extras = ["django"]}
requests = {version = "2.23.0", extras = ["security"]}
sentry-sdk = "0.14.2"
simplejson = "3.17.0"
whitenoise = "5.0.1"

[tool.poetry.dev-dependencies]
coverage = "^5.0.3"
django-debug-toolbar = "2.2"
freezegun = "0.3.15"
vcrpy = "4.0.2"
ipython = "^7.11.1"

[tool.isort]
line_length = 88
atomic = true
use_parentheses = true
include_trailing_comma = true
multi_line_output = 3
not_skip = "__init__.py"
filter_files = true
known_third_party = [
    "cache_memoize",
    "dotenv",
    "freezegun",
    "pycoingecko",
    "rest_framework",
    "sentry_sdk",
    "simplejson",
    "vcr",
]
known_django = "django"
skip_glob = [
    "java_wallet/models.py",
    "java_wallet/models_new.py",
    "java_wallet/migrations/0001_initial.py",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
    \.eggs
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | __pycache__
    | _build
    | build
    | dist
  )/
  | java_wallet/models.py
  | java_wallet/models_new.py
  | java_wallet/migrations/0001_initial.py
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

